Uporządkowanie kodu
