+++ blogApiTests/src/test/java/edu/iis/mto/blog/rest/test/CreateUserTest.java
@@ -23,4 +23,20 @@ public class CreateUserTest extends FunctionalTests {
                 .get("/blog/user/" + id);
     }
 
+    @Test
+    public void addUserWithUniqueMailShouldReturnCreateStatus() {
+        JSONObject jsonObj = new JSONObject().put("email", "adam@test.pl");
+        RestAssured.given().accept(ContentType.JSON).header("Content-Type", "application/json;charset=UTF-8")
+                .body(jsonObj.toString()).expect().log().all().statusCode(HttpStatus.SC_CREATED).when()
+                .post("/blog/user");
+    }
+
+    @Test
+    public void addUserWithoutUniqueMailShouldReturnConflictStatus() {
+        JSONObject jsonObj = new JSONObject().put("email", "john@domain.com");
+        RestAssured.given().accept(ContentType.JSON).header("Content-Type", "application/json;charset=UTF-8")
+                .body(jsonObj.toString()).expect().log().all().statusCode(HttpStatus.SC_CONFLICT).when()
+                .post("/blog/user");
+    }
+
 }
